Restructure project for Gradle
