+++ src/main/java/gitflow/GitflowConfigurable.java
