+++ src/main/java/gitflow/GitflowOptionsFactory.java
