+++ src/main/java/gitflow/GitflowState.java
