+++ src/main/java/gitflow/ui/GitflowCloseTaskPanel.form
