+++ src/main/java/gitflow/ui/GitflowCloseTaskPanel.java
