+++ src/main/java/gitflow/ui/GitflowOpenTaskPanel.form
