+++ src/main/java/gitflow/ui/GitflowOpenTaskPanel.java
