+++ src/main/java/gitflow/ui/GitflowTaskDialogPanelProvider.java
